// interfaces/user.interface.ts
export interface User {
  firts_name: string;
  last_name: string;
  country: string;
  dial_code: string;
  bird_day: string;
  phone: string;
  accept_send_marketing: boolean;
  accept_personal_data: boolean;
  accept_term: boolean;
  email: string;
  password: string;
  confirm_password: string;
}